[Refactor][Attention] Read dtype at forward for the varlen, sliding-window and DeepSeek ops
Eight more ops build their kernel on first forward, cached per dtype: the
three NSA ops, the DeepSeek sparse-MLA and MLA decode ops, GQA prefill
varlen, and both GQA sliding-window ops.

`GroupedQueryAttentionSlidingWindowFwdOp.total_memory` computed
`torch.tensor([], dtype=self.dtype).element_size()`. With no constructor
dtype, `dtype=None` makes that float32, so it silently returned twice the
real byte count. It now refuses to answer before the first forward rather
than guessing.

Three tests changed because the coupling they asserted is gone, not the
invariant underneath:

- an unsupported element type is still rejected, but the check moved with
  the dtype to first use, so the test feeds float32 tensors instead of
  naming float32 at construction;
- `total_memory` is asserted after a forward, and before one the test now
  requires the refusal;
- the dtype disagreement that still matters is between q, k and v, so the
  mismatch test makes k differ from q rather than making all three differ
  from a constructor argument.

Call sites needed all four forms: keyword, last-positional, an explicit
`params` dict inside the op, and a `**params` dict at the call.

+++ tests/ops/attention/test_deepseek_dsa_decode.py
@@ -80,7 +80,7 @@ def test_sparse_mla_decode(batch: int, heads: int, seq_len_q: int, seq_len_kv: i
         q_start_index_s, sm_scale=sm_scale, dtype=dtype)
     op = DeepSeekSparseAttentionDecodeWithKVCacheFwdOp(
         batch, heads, seq_len_q, seq_len_kv, dim, dim_tail, topk, stride_kv, heads_kv,
-        q_start_index_s, sm_scale=sm_scale, dtype=dtype, tune=tune)
+        q_start_index_s, sm_scale=sm_scale, tune=tune)
     test.check(op, *test.gen_inputs(), atol=3e-4, rtol=1e-5)
