+++ tests/ops/attention/test_deepseek_nsa_cmp.py
@@ -105,7 +105,7 @@ def test_nsa_cmp_fwd_varlen_op(
 
     op = NSACmpFwdVarlenOp(
         seq_num=seq_num, c_seq_len=c_seq_len, heads=heads, dim_k=dim_k, dim_v=dim_v, group=group,
-        scale=scale, bc=bc, bs=bs, dtype=dtype, accum_dtype=accum_dtype, tune=tune,
+        scale=scale, bc=bc, bs=bs, accum_dtype=accum_dtype, tune=tune,
         chunk_num=test.chunk_num)
     test.check(op, *inputs, atol=4e-3, rtol=1e-5)
